refactor: Move compiler, common, and macro tests to dedicated files
Complete test migration by moving remaining embedded tests:
- Compiler scanner tests (2 tests)
- Common module tests (56 tests from 3 files)
- Macro utility tests (4 tests)

All implementation files now contain only production code
