+++ src/common/tests/mod.rs
@@ -1,0 +1,4 @@
+mod macros;
+mod common_types;
+mod method_registry;
+mod string_similarity;
